# web/src/lib/api/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Athenaeum HTTP API",
    "description": "Session cookie (browser): sign in via POST /api/auth/login; cookies are sent automatically.\n\nHTTP Basic Auth: Authorization: Basic base64(username:password) — useful for OPDS clients and scripts.\n\nAPI key: Authorization: Bearer ath_\u003ctoken\u003e or X-API-Key: ath_\u003ctoken\u003e — create keys in Settings -\u003e API.\n\nJSON endpoints accept and return application/json unless noted.\n\nFile endpoints stream binary content with appropriate Content-Type and Accept-Ranges.\n\nMutating browser requests require the X-CSRF-Token header (not needed for Basic Auth or API keys).",
    "version": "dev"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
  "paths": {
    "/api/admin/pocketid": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Pocket ID connector settings",
        "description": "Auth: admin",
        "operationId": "GET__api_admin_pocketid",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Pocket ID connector",
        "description": "Auth: admin",
        "operationId": "PUT__api_admin_pocketid",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/pocketid/apply-oidc": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Apply Pocket ID issuer to OIDC config",
        "description": "Auth: admin",
        "operationId": "POST__api_admin_pocketid_apply_oidc",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/pocketid/test": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Test Pocket ID API key",
        "description": "Auth: admin",
        "operationId": "POST__api_admin_pocketid_test",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/server": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Server settings (metrics, proxy, CSP, CORS)",
        "description": "Auth: admin",
        "operationId": "GET__api_admin_server",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update server settings",
        "description": "Auth: admin",
        "operationId": "PUT__api_admin_server",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/auth/sessions": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Your active sessions",
        "description": "Auth: required",
        "operationId": "GET__api_auth_sessions",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/auth/sessions/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Revoke session",
        "description": "Auth: required",
        "operationId": "DELETE__api_auth_sessions__id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header",
        "description": "API key (ath_…)"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Username and password (OPDS and scripts)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key as Bearer token"
      },
      "sessionCookie": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie",
        "description": "Browser session after POST /api/auth/login"
      }
    }
  },
...
}